backend/processor/requirements.txt
# --- Web framework ---
Flask==3.0.3
gunicorn==22.0.0          # production WSGI server

# --- Google Cloud clients used by analyzer_main.py ---
google-cloud-core==2.4.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.19.0

# --- DSP / analysis libs your internal code is likely to use ---
numpy>=1.26.4
scipy>=1.13.0
librosa>=0.10.2
soundfile>=0.12.1
